Move payment-related components around
