+++ apps/maptio/src/app/modules/payment/payment.module.ts
@@ -7,9 +7,9 @@ import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { SanitizerModule } from '@maptio-shared/sanitizer.module';
 import { LoginModule } from '@maptio-login/login.module';
 
-import { CheckoutComponent } from './pages/checkout/checkout.page';
-import { PricingComponent } from './pages/pricing/pricing.page';
-import { PaymentPlanComponent } from './pages/pricing/payment-plan.component';
+import { CheckoutComponent } from './checkout/checkout.page';
+import { PricingComponent } from './pricing/pricing.page';
+import { PaymentPlanComponent } from './pricing/payment-plan.component';
 import { PaymentRoutingModule } from './payment.routing';
 
 @NgModule({
